pyproject.toml
[tool.poetry]
name = "ai-qna"
version = "0.1.0"
description = ""
authors = ["Connor (Cub) <connor@cubdigital.co.nz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.14"
python-dotenv = "^1.0.1"
openai = "^1.16.2"
chromadb = "^0.4.24"
tiktoken = "^0.6.0"
langchain-openai = "^0.1.1"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
